Adds support for python modules, splits out diff/tree functions.
