+++ autoload/gundo.py
@@ -9,8 +9,6 @@
 #
 # ============================================================================
 
-import difflib
-import itertools
 import re
 import sys
 import time
@@ -18,93 +16,20 @@ import tempfile
 
 try:
     import vim
+    # Add the library to the Python path.
+    for p in vim.eval("&runtimepath").split(','):
+        plugin_dir = os.path.join(p, "autoload")
+        if os.path.exists(os.path.join(plugin_dir, "mundo")):
+           if plugin_dir not in sys.path:
+              sys.path.append(plugin_dir)
+           break
 except:
     # vim isn't needed and isn't in the classpath when doing unit tests.
     pass
 
-# one line diff functions.#{{{
-def one_line_diff_str(before,after,mx=15,pre=2):
-    """
-    Return a summary of the differences between two strings, concatenated.
+from mundo.node import Nodes
+import mundo.util as util
 
-    Parameters:
-
-      before - string before.
-      after  - after string.
-      mx     - the max number of strings.
-      pre    - number of characters to show before diff (context)
-
-    Returns a string no longer than 'mx'.
-    """
-    old = one_line_diff(before,after)
-    result = ''
-    firstEl = True
-    # TODO instead of using +addition+ and -subtraction- it'd be nice to be able
-    # to highlight the change w/o requiring the +/- chars.
-    for v in old:
-        # if the first element doesn't have a change, then don't include it.
-        v = escape_returns(v)
-        if firstEl:
-            firstEl = False
-            # add in pre character context:
-            if not (v.startswith('+') or v.startswith('-')) and result == '':
-                v = v[-pre:]
-        # when we're going to be bigger than our max limit, lets ensure that the
-        # trailing +/- appears in the text:
-        if len(result) + len(v) > mx:
-            if v.startswith('+') or v.startswith('-'):
-                result += v[:mx - len(result) - 1]
-                result += v[0]
-            break
-        result += v
-    return result
-
-def escape_returns(result):
-    return result.replace('\n','\\n').replace('\r','\\r').replace('\t','\\t')
-
-def one_line_diff(before,after):
-  """
-  Return a summary of the differences between two arbitrary strings.
-
-  Returns a list of strings, summarizing all the changes.
-  """
-  s = difflib.SequenceMatcher(None,before,after)
-  results = []
-  for tag, i1, i2, j1, j2 in s.get_opcodes():
-    #print ("%7s a[%d:%d] (%s) b[%d:%d] (%s)" % (tag, i1, i2, before[i1:i2], j1, j2, after[j1:j2]))
-    if tag == 'equal':
-      _append_result(results,{
-        'equal': after[j1:j2]
-      })
-    if tag == 'insert':
-      _append_result(results,{
-        'plus': after[j1:j2]
-      })
-    elif tag == 'delete':
-      _append_result(results,{
-        'minus': before[i1:i2]
-      })
-    elif tag == 'replace':
-      _append_result(results,{
-        'minus': before[j1:j2],
-        'plus': after[j1:j2]
-      })
-  final_results = []
-  # finally, create a human readable string of information.
-  for v in results:
-    if 'minus' in v and 'plus' in v and len(v['minus']) > 0 and len(v['plus']) > 0:
-      final_results.append("-%s-+%s+"% (v['minus'],v['plus']))
-    elif 'minus' in v and len(v['minus']) > 0:
-      final_results.append("-%s-"% (v['minus']))
-    elif 'plus' in v and len(v['plus']) > 0:
-      final_results.append("+%s+"% (v['plus']))
-    elif 'equal' in v:
-      final_results.append("%s"% (v['equal']))
-  return final_results
-
-def _append_result(results,val):
-  results.append(val)
-#}}}
 # Mercurial's graphlog code --------------------------------------------------------#{{{
 def asciiedges(seen, rev, parents):
     """adds edge info to changelog DAG walk suitable for ascii()"""
@@ -306,7 +231,7 @@ def generate(dag, edgefn, current, verbose, num_header_lines, first_visible_line
         new_lines = ascii(state, 'C', char, [line], edgefn(seen, node, parents), verbose)
         line_number += len(new_lines)
         result.extend(new_lines)
-    _undo_to(current)
+    util._undo_to(current)
     return result
 
 # Mercurial age function -----------------------------------------------------------
@@ -374,22 +299,6 @@ def _check_sanity():
 
     return True
 
-def _goto_window_for_buffer(b):
-    w = int(vim.eval('bufwinnr(%d)' % int(b)))
-    vim.command('%dwincmd w' % int(w))
-
-def _goto_window_for_buffer_name(bn):
-    b = vim.eval('bufnr("%s")' % bn)
-    return _goto_window_for_buffer(b)
-
-def _undo_to(n):
-    n = int(n)
-    if n == 0:
-        vim.command('silent earlier %s' % (int(vim.eval('&undolevels')) + 1))
-    else:
-        vim.command('silent undo %d' % int(n))
-
-
 INLINE_HELP = '''\
 " Gundo (%d) - Press ? for Help:
 " %s/%s  - Next/Prev undo state.
@@ -406,193 +315,12 @@ INLINE_HELP = '''\
 '''
 
 #}}}
-# Python undo tree data structures and functions -----------------------------------#{{{
-class Node(object):
-    def __init__(self, n, parent, time, curhead, saved):
-        self.n = int(n)
-        self.parent = parent
-        self.children = []
-        self.curhead = curhead
-        self.saved = saved
-        self.time = time
-
-    def __repr__(self):
-        return "[n=%s,parent=%s,time=%s,curhead=%s,saved=%s]" % \
-            (self.n,self.parent,self.time,self.curhead,self.saved)
-
-class Nodes(object):
-    def __init__(self):
-        self._clear_cache()
-
-    def _clear_cache(self):
-        self.nodes_made = None
-        self.target_f = None
-        self.seq_last = None
-        self.lines = {}
-        self.diffs = {}
-        self.diff_has_oneline = {}
-
-    def _check_version_location(self):
-        _goto_window_for_buffer(vim.eval('g:gundo_target_n'))
-        target_f = vim.eval('g:gundo_target_f')
-        if target_f != self.target_f:
-            self._clear_cache()
-
-    def _make_nodes(self,alts, nodes, parent=None):
-        p = parent
-
-        for alt in alts:
-            if alt:
-                curhead = 'curhead' in alt
-                saved = 'save' in alt
-                node = Node(n=alt['seq'], parent=p, time=alt['time'], curhead=curhead, saved=saved)
-                nodes.append(node)
-                if alt.get('alt'):
-                    self._make_nodes(alt['alt'], nodes, p)
-                p = node
-
-    def make_nodes(self):
-        self._check_version_location()
-        target_f = vim.eval('g:gundo_target_f')
-        ut = vim.eval('undotree()')
-        entries = ut['entries']
-        seq_last = ut['seq_last']
-
-        # if the current seq_last and file are different, compute the new
-        # values:
-        if self.seq_last != seq_last:
-            root = Node(0, None, False, 0, 0)
-            nodes = []
-            # TODO only compute new values (not all values)
-            self._make_nodes(entries, nodes, root)
-            nodes.append(root)
-            nmap = dict((node.n, node) for node in nodes)
-
-            # cache values for later use
-            self.target_f = target_f
-            self.seq_last = seq_last
-            self.nodes_made = (nodes, nmap)
-
-        return self.nodes_made
-
-    def current(self):
-        """ Return the number of the current change. """
-        self._check_version_location()
-        nodes, nmap = self.make_nodes()
-        _curhead_l = list(itertools.dropwhile(lambda n: not n.curhead, nodes))
-        if _curhead_l:
-            current = _curhead_l[0].parent.n
-        else:
-            current = int(vim.eval('changenr()'))
-        return current
-
-    def _fmt_time(self,t):
-        return time.strftime('%Y-%m-%d %I:%M:%S %p', time.localtime(float(t)))
-
-    def _get_lines(self,node):
-        n = 0
-        if node:
-            n = node.n
-        if n not in self.lines:
-            _undo_to(n)
-            self.lines[n] = vim.current.buffer[:]
-        return self.lines[n]
-
-    def change_preview_diff(self,before,after):
-        self._check_version_location()
-        key = "%s-%s-cpd"%(before.n,after.n)
-        if key in self.diffs:
-            return self.diffs[key]
-
-        _goto_window_for_buffer(vim.eval('g:gundo_target_n'))
-        before_lines = self._get_lines(before)
-        after_lines = self._get_lines(after)
-
-        before_name = before.n or 'Original'
-        before_time = before.time and self._fmt_time(before.time) or ''
-        after_name = after.n or 'Original'
-        after_time = after.time and self._fmt_time(after.time) or ''
-
-        _undo_to(self.current())
-
-        self.diffs[key] = list(difflib.unified_diff(before_lines, after_lines,
-                                         before_name, after_name,
-                                         before_time, after_time))
-        return self.diffs[key]
-
-    def preview_diff(self, before, after, unified=True, inline=False):
-        """
-        Generate a diff comparing two versions of a file.
-
-        Parameters:
-
-          current - ?
-          before
-          after
-          unified - If True, generate a unified diff
-          inline - Generate a one line summary line.
-        """
-        self._check_version_location()
-        bn = 0
-        an = 0
-        if not after.n:    # we're at the original file
-            pass
-        elif not before.n: # we're at a pseudo-root state
-            an = after.n
-        else:
-            bn = before.n
-            an = after.n
-        key = "%s-%s-pd-%s"%(bn,an,unified)
-        needs_oneline = inline and key not in self.diff_has_oneline
-        if key in self.diffs and not needs_oneline:
-            return self.diffs[key]
-
-        if not after.n:    # we're at the original file
-            before_lines = []
-            after_lines = self._get_lines(None)
-
-            before_name = 'n/a'
-            before_time = ''
-            after_name = 'Original'
-            after_time = ''
-        elif not before.n: # we're at a pseudo-root state
-            before_lines = self._get_lines(None)
-            after_lines = self._get_lines(after)
-
-            before_name = 'Original'
-            before_time = ''
-            after_name = after.n
-            after_time = self._fmt_time(after.time)
-        else:
-            before_lines = self._get_lines(before)
-            after_lines = self._get_lines(after)
-
-            before_name = before.n
-            before_time = self._fmt_time(before.time)
-            after_name = after.n
-            after_time = self._fmt_time(after.time)
-
-        if unified:
-            self.diffs[key] = list(difflib.unified_diff(before_lines, after_lines,
-                                             before_name, after_name,
-                                             before_time, after_time))
-        elif inline:
-            maxwidth = int(vim.eval("col('$')"))
-            self.diffs[key] = one_line_diff_str('\n'.join(before_lines),'\n'.join(after_lines),maxwidth)
-            self.diff_has_oneline[key] = True
-        else:
-            self.diffs[key] = ""
-
-        return self.diffs[key]
-
-
-#}}}
 
 nodesData = Nodes()
 
 # Rendering utility functions
 def _output_preview_text(lines):
-    _goto_window_for_buffer_name('__Gundo_Preview__')
+    util._goto_window_for_buffer_name('__Gundo_Preview__')
     vim.command('setlocal modifiable')
     vim.current.buffer[:] = [line.rstrip() for line in lines]
     vim.command('setlocal nomodifiable')
@@ -670,12 +398,12 @@ def GundoRenderPreview():
     # Check that there's an undo state. There may not be if we're talking about
     # a buffer with no changes yet.
     if target_state == None:
-        _goto_window_for_buffer_name('__Gundo__')
+        util._goto_window_for_buffer_name('__Gundo__')
         return
     else:
         target_state = int(target_state)
 
-    _goto_window_for_buffer(vim.eval('g:gundo_target_n'))
+    util._goto_window_for_buffer(vim.eval('g:gundo_target_n'))
 
     nodes, nmap = nodesData.make_nodes()
 
@@ -685,11 +413,11 @@ def GundoRenderPreview():
     vim.command('call s:GundoOpenPreview()')
     _output_preview_text(nodesData.preview_diff(node_before, node_after))
 
-    _goto_window_for_buffer_name('__Gundo__')
+    util._goto_window_for_buffer_name('__Gundo__')
 
 def GundoGetTargetState():
     """ Get the current undo number that gundo is at.  """
-    _goto_window_for_buffer_name('__Gundo__')
+    util._goto_window_for_buffer_name('__Gundo__')
     target_line = vim.eval("getline('.')")
     return int(re.match('^.* \[([0-9]+)\] .*$',target_line).group(1))
 
@@ -803,12 +531,12 @@ def GundoMatch(down):
 
     # save the current window number (should be the navigation window)
     # then generate the undo nodes, and then go back to the current window.
-    _goto_window_for_buffer(vim.eval('g:gundo_target_n'))
+    util._goto_window_for_buffer(vim.eval('g:gundo_target_n'))
 
     nodes, nmap = nodesData.make_nodes()
     total = len(nodes) - 1
 
-    _goto_window_for_buffer_name('__Gundo__')
+    util._goto_window_for_buffer_name('__Gundo__')
     curline = int(vim.eval("line('.')"))
     gundo_node = GundoGetTargetState()
 
@@ -818,7 +546,7 @@ def GundoMatch(down):
         if down < 0:
             therange = range(gundo_node+1,total+1)
         for version in therange:
-            _goto_window_for_buffer_name('__Gundo__')
+            util._goto_window_for_buffer_name('__Gundo__')
             undochanges = nodesData.preview_diff(nmap[version].parent, nmap[version])
             # Look thru all of the changes, ignore the first two b/c those are the
             # diff timestamp fields (not relevent):
@@ -832,7 +560,7 @@ def GundoMatch(down):
             # found something, lets get out of here:
             if found_version != -1:
                 break
-    _goto_window_for_buffer_name('__Gundo__')
+    util._goto_window_for_buffer_name('__Gundo__')
     if found_version >= 0:
         GundoMove(found_version,1,False)
 
@@ -841,19 +569,19 @@ def GundoRenderPatchdiff():
     current file. """
     if GundoRenderChangePreview():
         # if there are no lines, do nothing (show a warning).
-        _goto_window_for_buffer_name('__Gundo_Preview__')
+        util._goto_window_for_buffer_name('__Gundo_Preview__')
         if vim.current.buffer[:] == ['']:
             # restore the cursor position before exiting.
-            _goto_window_for_buffer_name('__Gundo__')
+            util._goto_window_for_buffer_name('__Gundo__')
             vim.command('unsilent echo "No difference between current file and undo number!"')
             return False
 
         # quit out of gundo main screen
-        _goto_window_for_buffer_name('__Gundo__')
+        util._goto_window_for_buffer_name('__Gundo__')
         vim.command('quit')
 
         # save the __Gundo_Preview__ buffer to a temp file.
-        _goto_window_for_buffer_name('__Gundo_Preview__')
+        util._goto_window_for_buffer_name('__Gundo_Preview__')
         (handle,filename) = tempfile.mkstemp()
         vim.command('silent! w %s' % (filename))
         # exit the __Gundo_Preview__ window
@@ -873,12 +601,12 @@ def GundoGetChangesForLine():
     # Check that there's an undo state. There may not be if we're talking about
     # a buffer with no changes yet.
     if target_state == None:
-        _goto_window_for_buffer_name('__Gundo__')
+        util._goto_window_for_buffer_name('__Gundo__')
         return False
     else:
         target_state = int(target_state)
 
-    _goto_window_for_buffer(vim.eval('g:gundo_target_n'))
+    util._goto_window_for_buffer(vim.eval('g:gundo_target_n'))
 
     nodes, nmap = nodesData.make_nodes()
 
@@ -897,7 +625,7 @@ def GundoRenderChangePreview():
     vim.command('call s:GundoOpenPreview()')
     _output_preview_text(GundoGetChangesForLine())
 
-    _goto_window_for_buffer_name('__Gundo__')
+    util._goto_window_for_buffer_name('__Gundo__')
 
     return True
 
@@ -919,12 +647,12 @@ def GundoRevert():
     target_n = GundoGetTargetState()
     back = vim.eval('g:gundo_target_n')
 
-    _goto_window_for_buffer(back)
-    _undo_to(target_n)
+    util._goto_window_for_buffer(back)
+    util._undo_to(target_n)
 
     vim.command('GundoRenderGraph')
     if int(vim.eval('g:gundo_return_on_revert')):
-        _goto_window_for_buffer(back)
+        util._goto_window_for_buffer(back)
 
     if int(vim.eval('g:gundo_close_on_revert')):
         vim.command('GundoToggle')
@@ -939,7 +667,7 @@ def GundoPlayTo():
 
     vim.command('echo "%s"' % back)
 
-    _goto_window_for_buffer(back)
+    util._goto_window_for_buffer(back)
     normal('zR')
 
     nodes, nmap = nodesData.make_nodes()
@@ -977,10 +705,10 @@ def GundoPlayTo():
         return
 
     for node in branch:
-        _undo_to(node.n)
+        util._undo_to(node.n)
         vim.command('GundoRenderGraph')
         normal('zz')
-        _goto_window_for_buffer(back)
+        util._goto_window_for_buffer(back)
         vim.command('redraw')
         vim.command('sleep %dm' % delay)
