+++ autoload/mundo/util.py
@@ -1,0 +1,16 @@
+import vim
+
+def _goto_window_for_buffer(b):
+    w = int(vim.eval('bufwinnr(%d)' % int(b)))
+    vim.command('%dwincmd w' % int(w))
+
+def _goto_window_for_buffer_name(bn):
+    b = vim.eval('bufnr("%s")' % bn)
+    return _goto_window_for_buffer(b)
+
+def _undo_to(n):
+    n = int(n)
+    if n == 0:
+        vim.command('silent earlier %s' % (int(vim.eval('&undolevels')) + 1))
+    else:
+        vim.command('silent undo %d' % int(n))
